h2life: game of life over http2.0: step 1: move conway/game into a package so it can be reused
